fix: update security config
